[3.9][WIP] Fix latest merge

+++ lib/Alchemy/Phrasea/Controller/Admin/TaskManager.php
@@ -14,6 +14,8 @@ namespace Alchemy\Phrasea\Controller\Admin;
 use Alchemy\Phrasea\Exception\InvalidArgumentException;
 use Alchemy\Phrasea\Form\TaskForm;
 use Alchemy\Phrasea\Model\Entities\Task;
+use Alchemy\Phrasea\Exception\RuntimeException;
+use Alchemy\Phrasea\Exception\XMLParseErrorException;
 use Silex\Application;
 use Silex\ControllerProviderInterface;
 use Symfony\Component\HttpFoundation\Request;
@@ -130,6 +132,10 @@ class TaskManager implements ControllerProviderInterface
 
     public function getLiveInformation(Application $app, Request $request)
     {
+        if (false === $app['phraseanet.configuration']['main']['task-manager']['enabled']) {
+            throw new RuntimeException('The use of the task manager is disabled on this instance.');
+        }
+
         if ($request->getRequestFormat() !== "json") {
             $app->abort(406, 'Only JSON format is accepted.');
         }
@@ -147,6 +153,10 @@ class TaskManager implements ControllerProviderInterface
 
     public function getScheduler(Application $app, Request $request)
     {
+        if (false === $app['phraseanet.configuration']['main']['task-manager']['enabled']) {
+            throw new RuntimeException('The use of the task manager is disabled on this instance.');
+        }
+
         if ($request->getRequestFormat() !== "json") {
             $app->abort(406, 'Only JSON format is accepted.');
         }
@@ -239,6 +249,10 @@ class TaskManager implements ControllerProviderInterface
 
     public function postTaskDelete(Application $app, Request $request, Task $task)
     {
+        if (false === $app['phraseanet.configuration']['main']['task-manager']['enabled']) {
+            throw new RuntimeException('The use of the task manager is disabled on this instance.');
+        }
+
         $app['manipulator.task']->delete($task);
 
         return $app->redirectPath('admin_tasks_list');
@@ -246,6 +260,10 @@ class TaskManager implements ControllerProviderInterface
 
     public function postStartTask(Application $app, Request $request, Task $task)
     {
+        if (false === $app['phraseanet.configuration']['main']['task-manager']['enabled']) {
+            throw new RuntimeException('The use of the task manager is disabled on this instance.');
+        }
+
         $app['manipulator.task']->start($task);
 
         return $app->redirectPath('admin_tasks_list');
@@ -253,6 +271,10 @@ class TaskManager implements ControllerProviderInterface
 
     public function postStopTask(Application $app, Request $request, Task $task)
     {
+        if (false === $app['phraseanet.configuration']['main']['task-manager']['enabled']) {
+            throw new RuntimeException('The use of the task manager is disabled on this instance.');
+        }
+
         $app['manipulator.task']->stop($task);
 
         return $app->redirectPath('admin_tasks_list');
@@ -267,6 +289,10 @@ class TaskManager implements ControllerProviderInterface
 
     public function postSaveTask(Application $app, Request $request, Task $task)
     {
+        if (false === $app['phraseanet.configuration']['main']['task-manager']['enabled']) {
+            throw new RuntimeException('The use of the task manager is disabled on this instance.');
+        }
+
         if (!$this->doValidateXML($request->request->get('settings'))) {
             return $app->json(['success' => false, 'message' => sprintf('Unable to load XML %s', $request->request->get('xml'))]);
         }
@@ -304,6 +330,10 @@ class TaskManager implements ControllerProviderInterface
 
     public function getTask(Application $app, Request $request, Task $task)
     {
+        if (false === $app['phraseanet.configuration']['main']['task-manager']['enabled']) {
+            throw new RuntimeException('The use of the task manager is disabled on this instance.');
+        }
+
         if ('json' === $request->getContentType()) {
             return $app->json(array_replace([
                     'id' => $task->getId(),
@@ -330,6 +360,10 @@ class TaskManager implements ControllerProviderInterface
 
     public function validateXML(Application $app, Request $request)
     {
+        if (false === $app['phraseanet.configuration']['main']['task-manager']['enabled']) {
+            throw new RuntimeException('The use of the task manager is disabled on this instance.');
+        }
+        
         return $app->json(['success' => $this->doValidateXML($request->getContent())]);
     }
